+++ OggConverter/src/Misc/Logs.cs
@@ -107,7 +107,8 @@ namespace OggConverter
             Directory.CreateDirectory("log\\locale_errors");
             File.WriteAllText($"log\\locale_errors\\{date}.txt", errorMessage);
 
-            Form1.instance.Log(String.Format("An error has occured with current localisation. The problem has been saved into {0}.txt", date));
+            if (Form1.instance != null)
+                Form1.instance.Log(String.Format("An error has occured with current localisation. The problem has been saved into {0}.txt", date));
         }
 
         static string GetSystemInfo()